feat:remove context and use rtk query
