+++ src/components/search.tsx
@@ -1,40 +1,21 @@
 import React, { useEffect } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
-import {
-  setSearch,
-  selectSearch,
-  setData,
-  setMainPageLoading,
-  selectItemsPerPage,
-  selectPage,
-} from '../store';
+import { setSearch, selectSearch } from '../store';
 import { ErrorButton } from './errorbutton';
-import { api } from '../api';
 
 export function Search() {
   let searchForLS = '';
   const dispatch = useDispatch();
   const search: string = useSelector(selectSearch);
-  const itemsPerPage: number = useSelector(selectItemsPerPage);
-  const page: number = useSelector(selectPage);
-
-  const { data } = api.useGetDataQuery({
-    pageNumber: page,
-    itemsPerPage: itemsPerPage,
-    search: searchForLS,
-  });
 
   useEffect(() => {
-    dispatch(setMainPageLoading(true));
-    if (data) {
-      dispatch(setData(data.results));
-      dispatch(setMainPageLoading(false));
-    }
+    dispatch(setSearch(localStorage.getItem('search') || ''));
   }, []);
 
   const handleInputChange = (event: React.ChangeEvent<HTMLInputElement>) => {
     searchForLS = event.target.value;
     dispatch(setSearch(event.target.value));
+    localStorage.setItem('search', event.target.value);
   };
 
   const handleClick = () => {
